PathwayMaxMin

A function to obtain the hypothetical max and min activation status of selected pathway for a given scRNA seq data set

Usage

PathwayMaxMin(x, pathway, scale.data = TRUE)

Arguments

x

A matrix of genes x cells

pathway

A character string specifying the pathway name.

scale.data

A logical indicating whether to use scaled data (scale.data = TRUE) or normalized data. Default is TRUE.

Value

The hypothetical value for Pathway on and off (max and min value for features)

Examples

data(fake_test_matrix) # load the fake test data
PathwayMaxMin(fake_test_matrix, "Wnt", scale.data = TRUE)
#>         pathway.on pathway.off
#> Lgr5      1.851647  -11.600077
#> Rnf43     1.479722   -9.998238
#> Lrp5      1.880735   -7.935503
#> Lrp6      1.635447   -7.746165
#> Fzd6      1.873864   -8.215280
#> Ctnnb1    1.896552   -8.294299
#> Gsk3b     1.981532  -11.693745
#> Ccnd1     1.975098  -13.129679
#> Axin2     1.994353  -13.235483
#> Myc       1.685388  -11.679067
#> Lef1      1.596498  -10.771246
#> Tcf7      1.595631  -10.436295
#> Tcf7l1    1.497348   -9.843343
#> Tcf7l2    1.716832  -11.026494
#> Tle1      1.891611  -12.632374
#> Apc       1.944771   -8.883710
#> Csnk1a1   1.808021   -8.091854
#> Dvl1     -8.132137    2.135900