Example Processed Data for Plotting

A dataset generated by running PreparePlotData on fake_test_metadata and fake_final_mds. Contains processed data ready for visualization, with the following features:

Usage

data(fake_to_plot)

Format

A data frame with the following columns:

V1

A numerical or categorical value depending on the specific analysis

normalized

Normalized numerical value representing the cell's data

genotype

Group classification of the cell (e.g., "WT", "Mutant")

scale

Scaled data for visualization purposes

Examples

data(fake_to_plot)
head(fake_to_plot)
#>                  V1 normalized genotype       scale
#> Cell819  -1.3616977  0.8360230       WT -0.31471616
#> Cell124  -4.8660336  0.8146600       WT -1.09238288
#> Cell1363  3.2841257  0.8643448   Mutant  0.71626440
#> Cell1460  0.1235202  0.8450772   Mutant  0.01487683
#> Cell117  -1.4956377  0.8352065       WT -0.34443955
#> Cell1294  1.1383404  0.8512637   Mutant  0.24008124